scripts/requirements.txt
# 数据采集
tushare>=1.4.0
akshare>=1.12.0

# 工具
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# 定时任务
apscheduler>=3.10.0

# 未来扩展：Web展示
# fastapi>=0.100.0
# uvicorn>=0.23.0
# jinja2>=3.1.0
